requirements.txt
streamlit
pyrebase4
stmol
py3Dmol
biotite
ipython_genutils==0.1.0
requests